getWorkUnits now returns double instead of int.  removed cast.  helper method calculates double.

+++ source/org/miradi/objects/Indicator.java
@@ -135,7 +135,7 @@ public class Indicator extends BaseObject
 	}
 	
 	@Override
-	public int getWorkUnits(DateRange dateRangeToUse) throws Exception
+	public double getWorkUnits(DateRange dateRangeToUse) throws Exception
 	{
 		if (getMethodRefs().isEmpty())
 			return super.getWorkUnits(dateRangeToUse);
